Move editor actions to their own file
