+++ src/editor/actions/__tests__/editor-actions.test.js
@@ -1,9 +1,39 @@
 import configureMockStore from "redux-mock-store";
 import thunk from "redux-thunk";
-import { moveCursorToNextChunk, updateEditorCursor } from "../actions";
+import {
+  moveCursorToNextChunk,
+  toggleWrapInEditors,
+  updateEditorCursor,
+  updateEditorSelections,
+  updateIomdContent
+} from "../editor-actions";
 
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
+
+describe("basic editor actions", () => {
+  it("updateIomdContent works as expected", () => {
+    const iomd = "foo";
+    expect(updateIomdContent(iomd)).toEqual({
+      type: "UPDATE_IOMD_CONTENT",
+      iomd
+    });
+  });
+
+  it("toggleWrapInEditors works as expected", () => {
+    expect(toggleWrapInEditors()).toEqual({
+      type: "TOGGLE_WRAP_IN_EDITORS"
+    });
+  });
+
+  it("updateEditorSelections works as expected", () => {
+    const selections = [1, 2];
+    expect(updateEditorSelections(selections)).toEqual({
+      type: "UPDATE_SELECTIONS",
+      selections
+    });
+  });
+});
 
 describe("updateEditorCursor - returns correct action", () => {
   it("should create an action to update the cursor (NO force param)", () => {
